Process Comment, Push
Pushing. Now I'm going to totally mess up this entire folder. I explain in detail what the damage event should do.
Now it's all a matter of implementing it in.
I'm mostly worried about how we're going to store the max health inside the entity. There are not many ways to store persistent values inside of an entity. We're probably using Scoreboard.